feat: refactor routing structure and add AppRouter component
